LayoutTests/inspector/audit/audit-manager.html

<!doctype html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="./resources/audit-test-fixtures.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("AuditManager");

    suite.addTestCase({
        name: "Adding an AuditTestSuite",
        description: "AuditManager should have one instantiated AuditTestSuite.",
        async test(){
            let auditManager = new WI.AuditManager;

            InspectorTest.expectThat(!auditManager.testSuites.length, "AuditManager should have 0 testSuite.");

            InspectorTest.log("Adding an AuditTestSuite to AuditManager.");
            auditManager.addTestSuite(testSuiteFixture1);
            let testSuite = auditManager.testSuites[0];

            InspectorTest.expectEqual(auditManager.testSuites.length, 1, "AuditManager should have 1 test suite.");
            InspectorTest.expectEqual(testSuite.name, "FakeTestSuite1", "New test suite has the correct name.");
            InspectorTest.expectThat(testSuite instanceof WI.AuditTestSuite, "New test suite is of AuditTestSuite.");
        }
    });

    suite.addTestCase({
        name: "Adding a duplicating AuditTestSuite",
        description: "Should throw exception for duplicated test suite.",
        async test() {
            let auditManager = new WI.AuditManager;

            auditManager.addTestSuite(testSuiteFixture1);
            auditManager.addTestSuite(testSuiteFixture2);

            InspectorTest.expectException(() => {
            	auditManager.addTestSuite(testSuiteFixture1);
            });
        }
    });

    suite.addTestCase({
        name: "Perform tests by AuditTestSuite.",
        description: "Should produce report for AuditTestSuite.",
        async test() {
            let auditManager = new WI.AuditManager;

            InspectorTest.assert(!auditManager._reports.size, "auditManager has no reports.");

            let testSuite = new testSuiteFixture1;
            let testCaseNames = testSuite.testCases.map(testCase => {
                return testCase.name;
            });
            
            await auditManager.runAuditTestByRepresentedObject(testSuite);

            let auditReport = auditManager.reports[0];

            InspectorTest.expectThat(auditReport instanceof WI.AuditReport, "Receive a report that is of instance AuditReport."); 
            InspectorTest.expectThat(!auditReport._isWritable, "AuditReport is not writable");
            InspectorTest.expectEqual(auditReport.resultsData.length, 2, "There are two results in AuditReport.");
            
            for (let i = 0; i < auditReport.resultsData.length; i++) {
                let resultToTest = auditReport.resultsData[i];
                InspectorTest.expectThat(resultToTest instanceof WI.AuditResult, `auditResults ${i}  is an instance of AuditResult.`)
                InspectorTest.expectThat(testCaseNames.indexOf(resultToTest.name) >= 0, `auditReport ${i} is expected for test case ${resultToTest.name}.`);
            }
            InspectorTest.log("AuditReport is not writable.");
            InspectorTest.log("Attempting to add another AuditResult to AuditReport.");
            let additionalReport = auditReport.resultsData[1];
            auditReport.addResult(additionalReport);

            InspectorTest.expectEqual(auditReport.resultsData.length, 2, "AuditReport no longer accepts new AuditResults.");
            InspectorTest.expectEqual(auditReport.representedTestSuite.id, testSuite.id, "Report represents the expected AuditTestSuite.");
        }
    });


    suite.addTestCase({
        name: "Perform a test by AuditTestCase.",
        description: "Should produce report for AuditTestCase.",
        async test() {
            let auditManager = new WI.AuditManager;
            auditManager.addTestSuite(testSuiteFixture1);
            let testCase = auditManager.testSuites[0].testCases[0];

            await auditManager.runAuditTestByRepresentedObject(testCase);

            let auditReport = auditManager.reports[0];

            InspectorTest.expectThat(auditReport instanceof WI.AuditReport, "Receive a report that is of instance AuditReport.");
            InspectorTest.expectEqual(auditReport.representedTestCases[0], testCase, "AuditReport represents the expected AuditTestCase.");
        }
    });

    suite.addTestCase({
        name: "AuditReports are unique.",
        description: "No AuditReport should represent the same AuditTestCase/Suite",
        async test() {
            let auditManager = new WI.AuditManager;
            let testSuite = new testSuiteFixture1;

            InspectorTest.log("Only the most recent AuditReport for a case/suite is retained.");

            let results = [await auditManager.runAuditTestByRepresentedObject(testSuite), await auditManager.runAuditTestByRepresentedObject(testSuite)];

            InspectorTest.expectEqual(results[0].representedTestSuite, results[1].representedTestSuite, "The report represents the correct AuditTestSuite.");
        }
    });

    suite.addTestCase({
        name: "Get AuditReport by AuditTestCase/Suite id.",
        description: "Should return the correct AuditReport.",
        async test() {
            let auditManager = new WI.AuditManager;
            let testSuite = new testSuiteFixture1;
            let testCase = testSuite.testCases[0];

            InspectorTest.log("Running a test for an AuditTestSuite and an AuditTestCase.");
            let results = [await auditManager.runAuditTestByRepresentedObject(testSuite), await auditManager.runAuditTestByRepresentedObject(testCase)];

            let auditReportForTestSuite = auditManager.reportForId(testSuite.id);
            let auditReportForTestCase = auditManager.reportForId(testCase.id);

            InspectorTest.expectEqual(results[0], auditReportForTestSuite, "The report represents the correct AuditTestSuite.");
            InspectorTest.expectEqual(results[1], auditReportForTestCase, "The report represents the correct AuditTestCase.");
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test for the AuditManager Instantiation.</p>
</body>
</html>
