LayoutTests/inspector/audit/audit-manager-expected.txt

Test for the AuditManager Instantiation.


== Running test suite: AuditManager
-- Running test case: Adding an AuditTestSuite
PASS: AuditManager should have 0 testSuite.
Adding an AuditTestSuite to AuditManager.
PASS: AuditManager should have 1 test suite.
PASS: New test suite has the correct name.
PASS: New test suite is of AuditTestSuite.

-- Running test case: Adding a duplicating AuditTestSuite
PASS: Should produce an exception.
Error: class testSuiteFixture1 already exists.

-- Running test case: Perform tests by AuditTestSuite.
PASS: Receive a report that is of instance AuditReport.
PASS: AuditReport is not writable
PASS: There are two results in AuditReport.
PASS: auditResults 0  is an instance of AuditResult.
PASS: auditReport 0 is expected for test case fakeTest1.
PASS: auditResults 1  is an instance of AuditResult.
PASS: auditReport 1 is expected for test case fakeTest2.
AuditReport is not writable.
Attempting to add another AuditResult to AuditReport.
PASS: AuditReport no longer accepts new AuditResults.
PASS: Report represents the expected AuditTestSuite.

-- Running test case: Perform a test by AuditTestCase.
PASS: Receive a report that is of instance AuditReport.
PASS: AuditReport represents the expected AuditTestCase.

-- Running test case: AuditReports are unique.
Only the most recent AuditReport for a case/suite is retained.
PASS: The report represents the correct AuditTestSuite.

-- Running test case: Get AuditReport by AuditTestCase/Suite id.
Running a test for an AuditTestSuite and an AuditTestCase.
PASS: The report represents the correct AuditTestSuite.
PASS: The report represents the correct AuditTestCase.
